fix process query result mapping
This change is intended to fix the trasfomation of the query result into
an object which afects structs with complex composition.
Two tests have been added to reflect the scenarios to test_utils.py.
The approach to solve thee problem consists of two subsequent steps that
strip the message schema from subtypes and ofstruct details.
These subtypes and struct details were being ignored and caused the
issue in the first place.